// frontend/src/client/types.gen.ts
export type TeacherRead = {
    /**
     * Firstname
     */
    firstName: string;
    /**
     * Middlename
     */
    middleName: string | null;
    /**
     * Lastname
     */
    lastName: string;
    /**
     * Email
     */
    email: string | null;
    /**
     * Phone
     */
    phone: string | null;
    /**
     * Department
     */
    department: string;
    /**
     * Title
     */
    title: string;
    /**
     * Id
     */
    id: number;
    /**
     * Fullname
     */
    fullName: string;
    /**
     * Createdat
     */
    createdAt: string;
    /**
     * Updatedat
     */
    updatedAt: string;
};